+++ src/server/ports/mushroom/platform/index.js
@@ -5,3 +5,7 @@ export {
 export {
   createTelegramBotGatewayPort
 } from './bot-gateway-service.js';
+
+export {
+  createWikiServicePort
+} from './wiki-service.js';